Move `ProductFormSection.SettingsRow` extension to where the enum is defined.
